(CLOUD-281) Add support for managing templates as well as machines
Rather than add a additional type for templates this follows the API and
makes whether a machine is a template or not a boolean flag. This has a few
advantages:

* Less code (no new type)
* Simpler interface (no new type)
* Faster (because we can prefetch all machines in one go rather than
  having to make a request per type)
* Avoid path clashes (templates and machines are really the same thing,
  so can't occupy the same path)
* Removes the need for source_machine and template params, collapsing
  down to just source

While adjusting the acceptance tests to match the new interface I also:

* Set them all to create machines in the test folder
* Named all the machines CLOUD-* as per the delete spec
* Removes the need for the extra fixtures template

We should look at rationalising the acceptance tests with some helpers
but that was out of scope for this PR, which is already on the large
size.