(MAINT) speed up tests runs by limiting long running calls
We have a specific tests for deleteing a machine using ensure=absent.
Each of the runs used in the after step would do a prefetch, which takes
a while on a nested vsphere setup.

This commit takes the pattern from the machine_delete spec and reuses it
everywhere from the helper.