(CLOUD-290) Allow executing a command on the guest when creating a vm
This is a fairly low level interface, aimed at making most things
possible rather than a specific think lovely. It's possible/likely if
this sees usage we'll build higher level interfaces but I think for both
expediency and seeing what people actually do in the real world we
should start with this.