(MAINT) remove tests with race condition
I'm removing these two tests as they have a race condition, and aren't
actually testing behaviour exposed by the module.

Specifically they assume that the process started has finished by the
time we ask for the details of it. Given the action and the details are
run via VMware tools that's not always the case.

The module doesn't expose anything about the command run so I think it's
safe to remove these tests.