(CLOUD-355) Fix issue with integration tests failing first test of run
Due to PE-10757, when using the puppetserver, the first call to the
master from an agent will fail to load the module correctly. This also
needs documenting in the README, waiting on discussion in DOC-2115 as we
should be able to reference shared documentation here.

This commit therefore does several things:

* Adds a run of puppet agent after the module has been installed.
  This triggers the relevant population on the server
* Updates beaker to the latest version
* Adds specific el and debian pre-suite steps to install the gems
  (also see CLOUD-368 and CLOUD-366)
* Expands the checks to include additional platforms including debian
  and scientific linux
* Adds the ability to run the acceptance tests locally using Vagrant
* Fixes and expands the Rake tasks for running the integration tests
* Fixes the installation of nokogiri on RedHat by installing patch