(MAINT) Handle vms being deleted during prefetch
While investigating the CI failures I thought one of the causes might be
that the state of the cluster was changing during the run. This turned
out not to be the cause, but I was able to trigger a problem by:

* Running puppet resource
* While the above was running, deleting a VM

This happens because of the gap between us querying for all machines and
then processing the results, resulting in more queries. The first query
will return a reference to the machine, but when we come to process it
the API calls will fail because it no-longer exists. This could cause
transient failures in the real world, although it's not something we've
actually seen in practice.

The fix is to catch the relevant fake exception at the top level and to
ignore it. We already retry first to see if this is a transient failure,
which deals with the case of machines just starting up. This change
makes the code more robust in situations where machines are destroyed as
well.