(CLOUD-370) Change formulation for API queries
This commit changes how we call the API, and is based on research into
the extraConfig failures. I found this commit on a vmware project that
uses rbvmomi:

https://github.com/vmware/rvc/commit/ec900bf0df1871ac1d0b429e79d38f7629181925

This led me to this the comment in the rbvmomi source code this states that collect should
"Efficiently retrieve multiple properties from an object"

https://github.com/rlane/rbvmomi/blob/5dc0ca33165519a83f4eb8765835ac8f17306f84/lib/rbvmomi/vim/ManagedObject.rb#L27-L30

In the worst case this further reduces the number of separate API calls,
we no longer call summary and config separately.

Given I've been unable to trigger the failure we've seen in CI locally
this may also resolve that too, with this change we no-longer ask for config
but instead ask for config.extraConfig directly.