(CLOUD-370) Ignore machines which have just booted in prefetch
When machines boot they take a little time to enter a configured state.
Although we retry in cases where this happens quickly this can take
time.

This commit means that if machines don't enter a configured state
quickly they are ignored in this run. Note that I didn't find an
internal representation of this state in the API, just that the value of
config will be nil until vCenter does it's thing.