(CLOUD-392) Remove shell scripts for running integration tests
The current approach to requiring a mainly duplicate shell script (as
well as the configuration file) for running the integration tests
doesn't scale very well. It's also hard to discover and undocumented,
and while the current Rake task is discoverable it means we have two
entry points.

This commits tries to kill lots of code at the same time as providing a
single interface. This also makes it easier to add new configs.