(CLOUD-345) Performance improvements for prefetch
Prefecth previously triggered lots of queries for every VM. This could
add around a second per vm in some cases, leading to long run times.

This changes collects lots of the properties up-front, drastically
reducing the query time.

From basic testing this patch reduced the runtime per vm by 50%, and the
runtime per template by as much as 80%.

This commit also adds basic per-vm profiling to the debug output, which
is useful when looking for additional optimisations or slow queries.