(MAINT) Remove duplicate test cases with race condition
I think these tests exposed a race condition (in the tests) with the new
performance improvements. The functions here are also tested in the
accpetance tests and are not specific to a master/agent setup.

The usecase here, of a machine creating a VM from a template and then
immediately creating a template from that VM is also unlikely. The speed
of application is also only possible with direct automation like this -
which I think is what triggers the race condition.