(FM-4623) Fails to load some instances
The module fails to load instances where the resource pool is
associated with the host or a compute resource that is not
a base ResourcePool or a ClusterComputeResource. An update has been made
to add the base ComputeResource to the filter spec as well as
querying for the ComputeResource in the `resource_pool_from_machine_data`
method.

An additional nil check was added because in some scenarios the
machine data will not contain ClusterComputeResource or
ComputeResource.