(FM-6227) Add support for a datastore option
Prior to this commit, a new VM would be provisioned onto the same
datastore as the `source`.

This commit adds a new parameter `datastore` that will allow for
changing the destination datastore for the VM. The default `datastore`
is set to the first datastore available on the destination
`resource_pool`.